+++ routes/comments.js
@@ -1,0 +1,56 @@
+const express = require("express");
+const router = express.Router();
+const Campground = require("../models/campground");
+const Comment = require("../models/comment");
+
+
+// ===============
+// COMMENT ROUTES
+// ===============
+
+// NEW COMMENT
+router.get("/campgrounds/:id/comments/new", isLoggedIn, function(req, res){
+    // find campground by id
+    Campground.findById(req.params.id, function(err, campground){
+        if(err){
+            console.log(err);
+        } else {
+            // render the new comment form
+            res.render("comments/new", {campground: campground});
+        }
+    });
+});
+
+// CREATE COMMENT
+router.post("/campgrounds/:id/comments", isLoggedIn, function(req, res){
+    // find campground by id
+    Campground.findById(req.params.id, function(err, campground){
+        if(err){
+            console.log(err);
+            res.redirect("/campgrounds");
+        } else {
+          // create new comment
+          Comment.create(req.body.comment, function(err, comment){
+            if(err){
+                console.log(err);
+            } else {
+                // connect new comment to campground
+                campground.comments.push(comment);
+                campground.save();
+                // redirect to campground show page
+                res.redirect("/campgrounds/" + campground._id);
+            }
+          });
+        }
+    });
+});
+
+// CHECK FOR LOGGED IN
+function isLoggedIn(req, res, next){
+    if(req.isAuthenticated()){
+        return next();
+    }
+    res.redirect("/login");
+};
+
+module.exports = router;
